Polidea: reverted change for EnergyExpenditure data type and added a type parameter to the SHNSetResultListener
git-svn-id: https://redmine.natlab.research.philips.com/svnroot/moonshine/trunk/03_software/Source/Android/Moonshine/shinelib@4514 3c2db180-f856-4c6d-be1c-585a5d938589

+++ src/main/java/com/pins/philips/shinelib/datatypes/SHNDataEnergyExpenditure.java
@@ -4,13 +4,13 @@ package com.pins.philips.shinelib.datatypes;
  * Created by 310188215 on 13/05/15.
  */
 public class SHNDataEnergyExpenditure extends SHNData {
-    private final double energyExpenditure;
+    private final int energyExpenditure;
 
-    public SHNDataEnergyExpenditure(double energyExpenditure) {
+    public SHNDataEnergyExpenditure(int energyExpenditure) {
         this.energyExpenditure = energyExpenditure;
     }
 
-    public double getEnergyExpenditure() {
+    public int getEnergyExpenditure() {
         return energyExpenditure;
     }
